Restante do UpdaterReferee (prepare)
git-svn-id: http://roboime.com.br/svn/RoboCup/Projeto Computacional/RobotIME/trunk@275 536dbed5-76cd-454d-90a9-4f62b6fdd354

+++ Inteligencia/UpdateReferee.h
@@ -12,10 +12,10 @@ namespace Inteligencia {
 		unsigned char cmd_counter;     // increments each time new command is set
 		unsigned char goals_blue;      // current score for blue team
 		unsigned char goals_yellow;    // current score for yellow team
-		unsigned short time_remaining; // seconds remaining for current game stage (network byte order)
+		int time_remaining;			   // seconds remaining for current game stage (network byte order)
 	public:
 		UpdateReferee(void);
-		UpdateReferee(char cmd, unsigned char cmd_counter, unsigned char goals_blue, unsigned char goals_yellow, unsigned short time_remaining);
+		UpdateReferee(char cmd, unsigned char cmd_counter, unsigned char goals_blue, unsigned char goals_yellow, int time_remaining);
 		~UpdateReferee(void);
 		void apply(Updater*);//go through the updaters list and apply the update
 	};
